ai-service/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
drf-yasg==1.21.7

# Machine Learning
torch==2.2.0
numpy==1.26.3

# Production Server
gunicorn==21.2.0

# Development Tools
django-extensions==3.2.3
ipython==8.20.0
